♻️ [api] Move getUserInfo to pole emploi authentication service
