+++ api/lib/domain/services/authentication-service.js
@@ -1,5 +1,4 @@
 const encryptionService = require('./encryption-service');
-const tokenService = require('./token-service');
 
 async function getUserByUsernameAndPassword({ username, password, userRepository }) {
   const foundUser = await userRepository.getByUsernameOrEmailWithRolesAndPassword(username);
@@ -13,20 +12,6 @@ async function getUserByUsernameAndPassword({ username, password, userRepository
   return foundUser;
 }
 
-async function getPoleEmploiUserInfo(idToken) {
-  const { given_name, family_name, nonce, idIdentiteExterne } = await tokenService.extractPayloadFromPoleEmploiIdToken(
-    idToken
-  );
-
-  return {
-    firstName: given_name,
-    lastName: family_name,
-    externalIdentityId: idIdentiteExterne,
-    nonce,
-  };
-}
-
 module.exports = {
-  getPoleEmploiUserInfo,
   getUserByUsernameAndPassword,
 };
