+++ api/tests/unit/domain/usecases/authenticate-pole-emploi-user_test.js
@@ -13,7 +13,6 @@ const logger = require('../../../../lib/infrastructure/logger');
 const authenticatePoleEmploiUser = require('../../../../lib/domain/usecases/authenticate-pole-emploi-user');
 
 describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
-  let authenticationService;
   let tokenService;
   let poleEmploiAuthenticationService;
 
@@ -28,10 +27,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
 
     poleEmploiAuthenticationService = {
       exchangeCodeForTokens: sinon.stub(),
-    };
-
-    authenticationService = {
-      getPoleEmploiUserInfo: sinon.stub(),
+      getUserInfo: sinon.stub(),
     };
 
     tokenService = {
@@ -78,7 +74,6 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
         redirectUri: 'redirectUri',
         stateReceived,
         stateSent,
-        authenticationService,
         tokenService,
         authenticationMethodRepository,
         poleEmploiTokensRepository,
@@ -96,7 +91,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
   context('When user has an account', function () {
     it('should call authenticate pole emploi user with code and redirectUri parameters', async function () {
       // given
-      _fakePoleEmploiAPI({ authenticationService, poleEmploiAuthenticationService });
+      _fakePoleEmploiAPI({ poleEmploiAuthenticationService });
       tokenService.createAccessTokenForPoleEmploi.returns('access-token');
 
       // when
@@ -107,7 +102,6 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
         redirectUri: 'redirectUri',
         stateReceived: 'state',
         stateSent: 'state',
-        authenticationService,
         tokenService,
         poleEmploiAuthenticationService,
         authenticationMethodRepository,
@@ -124,7 +118,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
 
     it('should call get pole emploi user info with id token parameter', async function () {
       // given
-      _fakePoleEmploiAPI({ authenticationService, poleEmploiAuthenticationService });
+      _fakePoleEmploiAPI({ poleEmploiAuthenticationService });
       tokenService.createAccessTokenForPoleEmploi.returns('access-token');
 
       // when
@@ -135,7 +129,6 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
         redirectUri: 'redirectUri',
         stateReceived: 'state',
         stateSent: 'state',
-        authenticationService,
         tokenService,
         poleEmploiAuthenticationService,
         authenticationMethodRepository,
@@ -144,7 +137,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
       });
 
       // then
-      expect(authenticationService.getPoleEmploiUserInfo).to.have.been.calledWith('idToken');
+      expect(poleEmploiAuthenticationService.getUserInfo).to.have.been.calledWith('idToken');
     });
 
     it('should call tokenService createAccessTokenForPoleEmploi function with user id', async function () {
@@ -153,7 +146,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
       user.externalIdentityId = '094b83ac-2e20-4aa8-b438-0bc91748e4a6';
       tokenService.createAccessTokenForPoleEmploi.returns('access-token');
 
-      _fakePoleEmploiAPI({ authenticationService, poleEmploiAuthenticationService });
+      _fakePoleEmploiAPI({ poleEmploiAuthenticationService });
       userRepository.findByExternalIdentifier.resolves({ id: 1 });
 
       // when
@@ -164,7 +157,6 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
         redirectUri: 'redirectUri',
         stateReceived: 'state',
         stateSent: 'state',
-        authenticationService,
         tokenService,
         poleEmploiAuthenticationService,
         authenticationMethodRepository,
@@ -178,7 +170,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
 
     it('should return accessToken and idToken', async function () {
       // given
-      const { poleEmploiTokens } = _fakePoleEmploiAPI({ authenticationService, poleEmploiAuthenticationService });
+      const { poleEmploiTokens } = _fakePoleEmploiAPI({ poleEmploiAuthenticationService });
       const authenticatedUserId = 1;
       tokenService.createAccessTokenForPoleEmploi.withArgs(authenticatedUserId).returns('access-token');
 
@@ -190,7 +182,6 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
         redirectUri: 'redirectUri',
         stateReceived: 'state',
         stateSent: 'state',
-        authenticationService,
         tokenService,
         poleEmploiAuthenticationService,
         authenticationMethodRepository,
@@ -208,7 +199,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
 
     it('should save last logged at date', async function () {
       // given
-      _fakePoleEmploiAPI({ authenticationService, poleEmploiAuthenticationService });
+      _fakePoleEmploiAPI({ poleEmploiAuthenticationService });
       tokenService.createAccessTokenForPoleEmploi.returns('access-token');
 
       // when
@@ -219,7 +210,6 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
         redirectUri: 'redirectUri',
         stateReceived: 'state',
         stateSent: 'state',
-        authenticationService,
         tokenService,
         poleEmploiAuthenticationService,
         authenticationMethodRepository,
@@ -235,7 +225,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
       it('should call authentication repository updatePoleEmploiAuthenticationComplementByUserId function', async function () {
         // given
         userRepository.findByExternalIdentifier.resolves({ id: 1 });
-        const { poleEmploiTokens } = _fakePoleEmploiAPI({ authenticationService, poleEmploiAuthenticationService });
+        const { poleEmploiTokens } = _fakePoleEmploiAPI({ poleEmploiAuthenticationService });
         tokenService.createAccessTokenForPoleEmploi.returns('access-token');
 
         // when
@@ -246,7 +236,6 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
           redirectUri: 'redirectUri',
           stateReceived: 'state',
           stateSent: 'state',
-          authenticationService,
           tokenService,
           poleEmploiAuthenticationService,
           authenticationMethodRepository,
@@ -271,7 +260,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
       it('should also save last logged at date', async function () {
         // given
         userRepository.findByExternalIdentifier.resolves({ id: 123 });
-        _fakePoleEmploiAPI({ authenticationService, poleEmploiAuthenticationService });
+        _fakePoleEmploiAPI({ poleEmploiAuthenticationService });
         tokenService.createAccessTokenForPoleEmploi.returns('access-token');
 
         // when
@@ -282,7 +271,6 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
           redirectUri: 'redirectUri',
           stateReceived: 'state',
           stateSent: 'state',
-          authenticationService,
           tokenService,
           poleEmploiAuthenticationService,
           authenticationMethodRepository,
@@ -299,7 +287,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
       context('When the user does not have a pole emploi authentication method', function () {
         it('should call authentication method repository create function with pole emploi authentication method in domain transaction', async function () {
           // given
-          const { poleEmploiTokens } = _fakePoleEmploiAPI({ authenticationService, poleEmploiAuthenticationService });
+          const { poleEmploiTokens } = _fakePoleEmploiAPI({ poleEmploiAuthenticationService });
           userRepository.findByExternalIdentifier.resolves(null);
           tokenService.createAccessTokenForPoleEmploi.returns('access-token');
 
@@ -311,7 +299,6 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
             redirectUri: 'redirectUri',
             stateReceived: 'state',
             stateSent: 'state',
-            authenticationService,
             tokenService,
             poleEmploiAuthenticationService,
             authenticationMethodRepository,
@@ -339,7 +326,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
       context('When the user does have a pole emploi authentication method', function () {
         it('should call authentication repository updatePoleEmploiAuthenticationComplementByUserId function', async function () {
           // given
-          const { poleEmploiTokens } = _fakePoleEmploiAPI({ authenticationService, poleEmploiAuthenticationService });
+          const { poleEmploiTokens } = _fakePoleEmploiAPI({ poleEmploiAuthenticationService });
           authenticationMethodRepository.findOneByUserIdAndIdentityProvider.resolves(
             domainBuilder.buildAuthenticationMethod.withPoleEmploiAsIdentityProvider({
               externalIdentifier: '094b83ac-2e20-4aa8-b438-0bc91748e4a6',
@@ -355,7 +342,6 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
             redirectUri: 'redirectUri',
             stateReceived: 'state',
             stateSent: 'state',
-            authenticationService,
             tokenService,
             poleEmploiAuthenticationService,
             authenticationMethodRepository,
@@ -379,7 +365,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
 
         it('should throw an UnexpectedUserAccountError error if the external identifier does not match the one in the pole emploi id token', async function () {
           // given
-          _fakePoleEmploiAPI({ authenticationService, poleEmploiAuthenticationService });
+          _fakePoleEmploiAPI({ poleEmploiAuthenticationService });
 
           authenticationMethodRepository.findOneByUserIdAndIdentityProvider.resolves(
             domainBuilder.buildAuthenticationMethod.withPoleEmploiAsIdentityProvider({
@@ -395,7 +381,6 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
             redirectUri: 'redirectUri',
             stateReceived: 'state',
             stateSent: 'state',
-            authenticationService,
             tokenService,
             poleEmploiAuthenticationService,
             authenticationMethodRepository,
@@ -413,7 +398,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
   context('When user has no account', function () {
     it('should call poleEmploiTokens repository save method', async function () {
       // given
-      const { poleEmploiTokens } = _fakePoleEmploiAPI({ authenticationService, poleEmploiAuthenticationService });
+      const { poleEmploiTokens } = _fakePoleEmploiAPI({ poleEmploiAuthenticationService });
       const key = 'aaa-bbb-ccc';
       poleEmploiTokensRepository.save.resolves(key);
       userRepository.findByExternalIdentifier.resolves(null);
@@ -426,7 +411,6 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
         redirectUri: 'redirectUri',
         stateReceived: 'state',
         stateSent: 'state',
-        authenticationService,
         tokenService,
         poleEmploiAuthenticationService,
         authenticationMethodRepository,
@@ -441,7 +425,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
     it('should return an authenticationKey', async function () {
       // given
       const key = 'aaa-bbb-ccc';
-      _fakePoleEmploiAPI({ authenticationService, poleEmploiAuthenticationService });
+      _fakePoleEmploiAPI({ poleEmploiAuthenticationService });
       poleEmploiTokensRepository.save.resolves(key);
       userRepository.findByExternalIdentifier.resolves(null);
 
@@ -453,7 +437,6 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
         redirectUri: 'redirectUri',
         stateReceived: 'state',
         stateSent: 'state',
-        authenticationService,
         tokenService,
         poleEmploiAuthenticationService,
         authenticationMethodRepository,
@@ -467,7 +450,7 @@ describe('Unit | UseCase | authenticate-pole-emploi-user', function () {
   });
 });
 
-function _fakePoleEmploiAPI({ authenticationService, poleEmploiAuthenticationService }) {
+function _fakePoleEmploiAPI({ poleEmploiAuthenticationService }) {
   const poleEmploiTokens = new PoleEmploiTokens({
     accessToken: 'poleEmploiAccessToken',
     expiresIn: 60,
@@ -481,7 +464,7 @@ function _fakePoleEmploiAPI({ authenticationService, poleEmploiAuthenticationSer
   };
 
   poleEmploiAuthenticationService.exchangeCodeForTokens.resolves(poleEmploiTokens);
-  authenticationService.getPoleEmploiUserInfo.resolves(userInfo);
+  poleEmploiAuthenticationService.getUserInfo.resolves(userInfo);
 
   return { poleEmploiTokens };
 }
